+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 # Changelog
 
-TODO: You can follow this format for your changelog: <https://keepachangelog.com/en/1.1.0/>
+## 0.1.1
+ - Updated changelog from the mod base template, lol.
+ - Added new setting Quota Modifier, which allows you to modify the quota based on the number of players in the lobby.
+	- There are 3 potential settings:
+	    - None: No modifications will be applied.
+		- OnlyScaleUp: Modifications will only be applied when there is more than 6 players.
+		- AlwaysScale: Quota will always be scaled based on the number of players in the lobby, resulting in easier quotas when there is less than 6 players.   
+
+## 0.1.0
+ - Initial release.